add changes

+++ src/main/java/com/oliva/marc/model/entity/Empowered.java
@@ -14,37 +14,37 @@ import javax.persistence.Table;
 public class Empowered implements Serializable {
 
 	private static final long serialVersionUID = 1L;
-	
+
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
-	
+	private Long id;
+
 	@Column(name = "name", length = 20, nullable = false)
 	private String name;
-	
+
 	@Column(name = "paternal_lastname", length = 20, nullable = false)
 	private String paternalLastname;
-	
+
 	@Column(name = "maternal_lastname", length = 20, nullable = false)
 	private String maternalLastname;
-	
+
 	@Column(name = "birthdate", nullable = false)
 	private String birthdate;
-	
+
 	@Column(name = "email", length = 50, nullable = false)
 	private String email;
-	
+
 	@Column(name = "dni", precision = 8, nullable = false)
 	private Integer dni;
-	
+
 	@Column(name = "phone", precision = 9, nullable = false)
 	private Integer phone;
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
