+++ src/main/java/com/oliva/marc/model/entity/Teacher.java
@@ -19,38 +19,38 @@ import io.swagger.annotations.ApiModel;
 public class Teacher implements Serializable {
 
 	private static final long serialVersionUID = 1L;
-	
+
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
-	
+	private Long id;
+
 	@Column(name = "name", length = 20, nullable = false)
 	private String name;
-	
+
 	@Column(name = "paternal_lastname", length = 20, nullable = false)
 	private String paternalLastname;
-	
+
 	@Column(name = "maternal_lastname", length = 20, nullable = false)
 	private String maternalLastname;
-	
+
 	@Column(name = "email", length = 50, nullable = false)
 	private String email;
-	
+
 	@Column(name = "dni", precision = 8, nullable = false)
 	private Integer dni;
-	
+
 	@Column(name = "phone", precision = 9, nullable = false)
 	private Integer phone;
-	
+
 	@ManyToOne
 	@JoinColumn(name = "education_degree_id", nullable = false)
 	private EducationDegree educationDegree;
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
